requirements.txt
# Python 3.8+ compatible versions
openai>=1.0.0,<2.0.0
psycopg2-binary>=2.8.6
PyMySQL>=1.0.0
python-dotenv>=0.19.0
numpy>=1.19.0
cryptography>=3.4.0
